minor script changes
