+++ service/create.html
@@ -89,8 +89,9 @@
         <script type="text/javascript" src="js/utilities/loader.js"></script>
         <script type="text/javascript" src="js/utilities/mapper.js"></script>
         <script type="text/javascript" src="js/utilities/namer.js"></script>
-        <script type="text/javascript" src="js/model/model.js"></script>
         <script type="text/javascript" src="js/utilities/fasteditor.js"></script>
+        <script type="text/javascript" src="js/utilities/misc.js"></script>
+        <script type="text/javascript" src="js/model.js"></script>
         <script type="text/javascript" src="js/api.js"></script>
         <script type="text/javascript" src="js/status.js"></script>
         <script type="text/javascript" src="js/create.js"></script>
