+++ service/index.html
@@ -64,9 +64,9 @@
         <script type="text/javascript" src="js/utilities/loader.js"></script>
         <script type="text/javascript" src="js/utilities/mapper.js"></script>
         <script type="text/javascript" src="js/utilities/misc.js"></script>
-        <script type="text/javascript" src="js/experimental/model.js"></script>
-        <script type="text/javascript" src="js/experimental/api.js"></script>
-        <script type="text/javascript" src="js/experimental/status.js"></script>
+        <script type="text/javascript" src="js/model.js"></script>
+        <script type="text/javascript" src="js/api.js"></script>
+        <script type="text/javascript" src="js/status.js"></script>
         <script type="text/javascript" src="js/jobs.js"></script>
     </body>
 </html>
