+++ service/log.html
@@ -56,6 +56,7 @@
         <script type="text/javascript" src="js/utilities/loader.js"></script>
         <script type="text/javascript" src="js/utilities/mapper.js"></script>
         <script type="text/javascript" src="js/utilities/misc.js"></script>
+        <script type="text/javascript" src="js/model.js"></script>
         <script type="text/javascript" src="js/api.js"></script>
         <script type="text/javascript" src="js/status.js"></script>
         <script type="text/javascript" src="js/log.js"></script>
