+++ service/timetable.html
@@ -61,6 +61,7 @@
         <script type="text/javascript" src="js/utilities/loader.js"></script>
         <script type="text/javascript" src="js/utilities/mapper.js"></script>
         <script type="text/javascript" src="js/utilities/misc.js"></script>
+        <script type="text/javascript" src="js/model.js"></script>
         <script type="text/javascript" src="js/api.js"></script>
         <script type="text/javascript" src="js/status.js"></script>
         <script type="text/javascript" src="js/timetabledetails.js"></script>
