+++ service/timetablecreate.html
@@ -187,7 +187,7 @@
         <script type="text/javascript" src="js/utilities/fasteditor.js"></script>
         <script type="text/javascript" src="js/utilities/misc.js"></script>
         <script type="text/javascript" src="js/utilities/namer.js"></script>
-        <script type="text/javascript" src="js/model/model.js"></script>
+        <script type="text/javascript" src="js/model.js"></script>
         <script type="text/javascript" src="js/api.js"></script>
         <script type="text/javascript" src="js/status.js"></script>
         <script type="text/javascript" src="js/timetablecreate.js"></script>
